Remove Matchmaking Scene Graph
