+++ build-win10/Sweetspace/Sweetspace.vcxproj
@@ -210,9 +210,8 @@
     <ClInclude Include="..\..\source\libraries\easywsclient.hpp" />
     <ClInclude Include="..\..\source\LoadingMode.h" />
     <ClInclude Include="..\..\source\MagicInternetBox.h" />
+    <ClInclude Include="..\..\source\MainMenuMode.h" />
     <ClInclude Include="..\..\source\PlayerDonutModel.h" />
-    <ClInclude Include="..\..\source\MatchmakingGraphRoot.h" />
-    <ClInclude Include="..\..\source\MatchmakingMode.h" />
     <ClInclude Include="..\..\source\ShipModel.h" />
     <ClInclude Include="..\..\source\Sweetspace.h" />
     <ClInclude Include="..\..\source\Tween.h" />
@@ -238,9 +237,8 @@
     <ClCompile Include="..\..\source\GameMode.cpp" />
     <ClCompile Include="..\..\source\InputController.cpp" />
     <ClCompile Include="..\..\source\LoadingMode.cpp" />
+    <ClCompile Include="..\..\source\MainMenuMode.cpp" />
     <ClCompile Include="..\..\source\PlayerDonutModel.cpp" />
-    <ClCompile Include="..\..\source\MatchmakingGraphRoot.cpp" />
-    <ClCompile Include="..\..\source\MatchmakingMode.cpp" />
     <ClCompile Include="..\..\source\ShipModel.cpp" />
     <ClCompile Include="..\..\source\Sweetspace.cpp" />
     <ClCompile Include="..\..\source\Tween.cpp" />
