+++ build-win10/Sweetspace/Sweetspace.vcxproj.filters
@@ -93,12 +93,6 @@
     <ClInclude Include="..\..\source\DoorNode.h">
       <Filter>Header Files\SceneGraph</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\source\MatchmakingMode.h">
-      <Filter>Header Files\Modes</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\source\MatchmakingGraphRoot.h">
-      <Filter>Header Files\SceneGraph</Filter>
-    </ClInclude>
     <ClInclude Include="..\..\source\GLaDOS.h">
       <Filter>Header Files\Controllers</Filter>
     </ClInclude>
@@ -134,6 +128,9 @@
     </ClInclude>
     <ClInclude Include="..\..\source\ButtonNode.h">
       <Filter>Header Files\SceneGraph</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\source\MainMenuMode.h">
+      <Filter>Header Files\Modes</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
@@ -185,9 +182,6 @@
     <ClCompile Include="..\..\source\DoorNode.cpp">
       <Filter>Source Files\SceneGraph</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\source\MatchmakingMode.cpp">
-      <Filter>Source Files\Modes</Filter>
-    </ClCompile>
     <ClCompile Include="..\..\source\GLaDOS.cpp">
       <Filter>Source Files\Controllers</Filter>
     </ClCompile>
@@ -203,14 +197,14 @@
     <ClCompile Include="..\..\source\ButtonManager.cpp">
       <Filter>Source Files\Helpers</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\source\MatchmakingGraphRoot.cpp">
-      <Filter>Source Files\SceneGraph</Filter>
-    </ClCompile>
     <ClCompile Include="..\..\source\ButtonNode.cpp">
       <Filter>Source Files\SceneGraph</Filter>
     </ClCompile>
     <ClCompile Include="..\..\source\ButtonModel.cpp">
       <Filter>Source Files\Models</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\source\MainMenuMode.cpp">
+      <Filter>Source Files\Modes</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
